Updated project to target .NET 10.0 and upgraded package references; added settings for permissions and improved type safety in various components.

+++ Demizon.Core/Services/FileUpload/FileUploadService.cs
@@ -55,7 +55,7 @@ public class FileUploadService(IOptionsSnapshot<UploadSettings> uploadSettings) 
                 continue;
             }
 
-            image.Resize(resizeDimensions.Width, resizeDimensions.Height);
+            image.Resize((uint)resizeDimensions.Width, (uint)resizeDimensions.Height);
             string resizedFileName = resizeName.ToLower() + "_" + fileName;
             string resizedFile = Path.GetDirectoryName(fileUri.AbsolutePath) + "/" + resizedFileName;
             image.Write(resizedFile);
